tests: add functional tests for metrics_show
